requirements.txt
streamlit
numpy
Pillow
opencv-python
imgaug
opencv-python